Multiple Hangar Types
Allows using/defining different hangar types to host just 1 craft each.

A new key ("hangarType") is added for both crafts & facilities
rulesets (YAML) so the value for the craft must match the value at the
hangar to be allocated inside.

If that key is not defined, a default value (-1) is assigned, so both
crafts and hangar facilities behave as in original OXCE.

Added a "MOD Advanced Option" to show hangar types in crafts' UEFOPEDIA
entries. Fof Facilities' UEFOPEDIA entries the modder should as the text
describing the hangar type.

Added "hangar Type" un "Stats for Nerds" screen for both crafts and
facilities.