Multiple Crafts x Hangar
Allows using more than 1 craft allocated in hangars. All hangars can
allocate any type of craft.

For hangars with more than 1 craft, a new key ("craftSlots") is added
to the facilities ruleset (YAML) in order to define so many 'position
 vectors' [x,y,z] as crafts are allowed:

    crafts: 2
    craftSlots: # mandatory if crafts > 1; optional otherwise
      - [11, -11, 0]  # Distance [x,y] to the
      - [-11, 11, 0]  # central point

z value is not used.

If that key is not defined (and crafts=1), a default slot ([0, 0, 0])
is assigned, so it behaves as in "original" OXCE.

Adjusted "damageFacility", "destroyfacility"&"removeCraft" at Base.cpp
to consider a vector of crafts instead a single one (so, more than one
craft and their allocated items and soldiers are considered)